actiondb-parser: move 'file' submodule's contents into mod.rs
The original code produced a warning.
