--- actiondb-parser/actiondb/src/matcher/pattern/file/file.rs
@@ -1,11 +1,0 @@
-use matcher::pattern::Pattern;
-
-pub struct PatternFile {
-    pub patterns: Vec<Pattern>,
-}
-
-impl PatternFile {
-    pub fn patterns(&self) -> &Vec<Pattern> {
-        &self.patterns
-    }
-}
