+++ actiondb-parser/actiondb/src/matcher/pattern/file/mod.rs
@@ -1,3 +1,14 @@
 mod deser;
-mod file;
 mod iter;
+
+use matcher::pattern::Pattern;
+
+pub struct PatternFile {
+    pub patterns: Vec<Pattern>,
+}
+
+impl PatternFile {
+    pub fn patterns(&self) -> &Vec<Pattern> {
+        &self.patterns
+    }
+}
